% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_results.R
\name{create_prediction_output}
\alias{create_prediction_output}
\title{Create a beautiful output for model predictions}
\usage{
create_prediction_output(subgroups)
}
\arguments{
\item{subgroups}{List of subgroups from analyze_entire_series}
}
\value{
A formatted string with the prediction results
}
\description{
Create a beautiful output for model predictions
}
